Module hacks: Add bindings to make POSIX and Socket modules work
Various bits of the POSIX functions and types live in libnx.

I'm wondering whether I should be creating my own versions of builtin
modules and putting them in the `modules` folder instead of using this
patch... It would be more self-contained, but the downside would be that
they'd be frozen in time.